ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
ultralytics==8.2.0
torch==2.3.0
torchvision==0.18.0
supervision==0.21.0
Pillow==10.3.0
opencv-python==4.10.0.84
numpy==1.26.4
python-socketio==5.11.0
aiofiles==23.2.1
zipfile36==0.1.3
transformers==4.38.2
pycocotools>=2.0.7
timm>=0.9.16
yapf==0.40.2
addict==2.4.0
